move getUserIdForVaulting to Config
